+++ src/module/data/actor/_types.d.ts
@@ -67,6 +67,9 @@ declare module "./base.mjs" {
     }
     potency: {
       bonuses: number;
+      weak: number;
+      average: number;
+      strong: number;
     }
   }
 }
@@ -90,12 +93,6 @@ declare module "./hero.mjs" {
       wealth: number;
       skills: Set<string>;
       preferredKit: string;
-    }
-    potency: {
-      bonuses: number;
-      weak: number;
-      average: number;
-      strong: number;
     }
     biography: Biography & {
       age: string;